bot/pyproject.toml
[project]
name = "riftteam-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.4",
    "aiohttp>=3.11",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio"]
